refactored reddit stat collector and update the sample config
