+++ commstats/commstat.go
@@ -72,68 +72,5 @@ func (c *Collector) Run(ctx context.Context) {
 	// github
 	go c.startGithubCollector(ctx)
 
-	err := c.collectAndStore(ctx)
-	app.ReleaseForNewModule()
-	if err != nil {
-		log.Errorf("Could not start collection: %v", err)
-		return
-	}
-
-	ticker := time.NewTicker(time.Duration(c.options.RedditStatInterval) * time.Second)
-	defer ticker.Stop()
-
-	for {
-		select {
-		case <-ctx.Done():
-			log.Infof("Shutting down community stats collector")
-			return
-		case <-ticker.C:
-			// continually check the state of the app until its free to run this module
-			for {
-				if app.MarkBusyIfFree() {
-					break
-				}
-			}
-			log.Info("Fetching community stats...")
-			err := c.collectAndStore(ctx)
-			app.ReleaseForNewModule()
-			if err != nil {
-				return
-			}
-		}
-	}
-}
-
-func (c *Collector) collectAndStore(ctx context.Context) error {
-	if ctx.Err() != nil {
-		return ctx.Err()
-	}
-
-	for _, subreddit := range c.options.Subreddit {
-		// reddit
-		resp := new(RedditResponse)
-		resp, err := c.fetchRedditStat(ctx, subreddit)
-		for retry := 0; err != nil; retry++ {
-			if retry == retryLimit {
-				return err
-			}
-			log.Warn(err)
-			resp, err = c.fetchRedditStat(ctx, subreddit)
-		}
-
-		err = c.dataStore.StoreRedditStat(ctx, Reddit{
-			Date:           time.Now().UTC(),
-			Subscribers:    resp.Data.Subscribers,
-			AccountsActive: resp.Data.AccountsActive,
-			Subreddit:      subreddit,
-		})
-		if err != nil {
-			log.Error("Unable to save reddit stat, %s", err.Error())
-			return err
-		}
-		log.Infof("New Reddit stat collected for %s at %s, Subscribers  %d, Active Users %d", subreddit,
-			time.Now().Format(dateMiliTemplate), resp.Data.Subscribers, resp.Data.AccountsActive)
-	}
-
-	return nil
+	go c.startRedditCollector(ctx)
 }
