+++ commstats/reddit.go
@@ -2,9 +2,15 @@ package commstats
 
 import (
 	"context"
+	"database/sql"
 	"encoding/json"
 	"fmt"
 	"net/http"
+	"time"
+
+	"github.com/raedahgroup/dcrextdata/app"
+	"github.com/raedahgroup/dcrextdata/app/helpers"
+	"github.com/raedahgroup/dcrextdata/postgres/models"
 )
 
 const (
@@ -17,16 +23,80 @@ func Subreddits() []string {
 	return subreddits
 }
 
-var twitterHandles []string
+func (c *Collector) startRedditCollector(ctx context.Context) {
+	var lastCollectionDate time.Time
+	err := c.dataStore.LastEntry(ctx, models.TableNames.Reddit, &lastCollectionDate)
+	if err != nil && err != sql.ErrNoRows {
+		log.Errorf("Cannot fetch last Reddit entry time, %s", err.Error())
+		return
+	}
 
-func TwitterHandles() []string {
-	return twitterHandles
+	secondsPassed := time.Since(lastCollectionDate)
+	period := time.Duration(c.options.TwitterStatInterval) * time.Minute
+
+	if secondsPassed < period {
+		timeLeft := period - secondsPassed
+		log.Infof("Fetching Reddit stats every %dm, collected %s ago, will fetch in %s.", period/time.Minute, helpers.DurationToString(secondsPassed),
+			helpers.DurationToString(timeLeft))
+
+		time.Sleep(timeLeft)
+	}
+
+	registerStarter := func() {
+		// continually check the state of the app until its free to run this module
+		for {
+			if app.MarkBusyIfFree() {
+				break
+			}
+		}
+	}
+
+	registerStarter()
+	c.collectAndStoreRedditStat(ctx)
+	app.ReleaseForNewModule()
+
+	ticker := time.NewTicker(time.Duration(c.options.RedditStatInterval) * time.Minute)
+	for {
+		select {
+		case <-ctx.Done():
+			return
+		case <-ticker.C:
+			registerStarter()
+			c.collectAndStoreRedditStat(ctx)
+			app.ReleaseForNewModule()
+		}
+	}
 }
 
-var repositories []string
+func (c *Collector) collectAndStoreRedditStat(ctx context.Context) {
+	log.Info("Starting Reddit stats collection cycle")
 
-func Repositories() []string {
-	return repositories
+	for _, subreddit := range c.options.Subreddit {
+		// reddit
+		resp := new(RedditResponse)
+		resp, err := c.fetchRedditStat(ctx, subreddit)
+		for retry := 0; err != nil; retry++ {
+			if retry == retryLimit {
+				log.Error(err)
+				return
+			}
+			log.Warn(err)
+			resp, err = c.fetchRedditStat(ctx, subreddit)
+		}
+
+		err = c.dataStore.StoreRedditStat(ctx, Reddit{
+			Date:           time.Now().UTC(),
+			Subscribers:    resp.Data.Subscribers,
+			AccountsActive: resp.Data.AccountsActive,
+			Subreddit:      subreddit,
+		})
+		if err != nil {
+			log.Error("Unable to save reddit stat, %s", err.Error())
+			return
+		}
+		log.Infof("New Reddit stat collected for %s at %s, Subscribers  %d, Active Users %d", subreddit,
+			time.Now().Format(dateMiliTemplate), resp.Data.Subscribers, resp.Data.AccountsActive)
+	}
 }
 
 func (c *Collector) fetchRedditStat(ctx context.Context, subreddit string) (response *RedditResponse, err error) {
