+++ sample-dcrextdata.conf
@@ -53,14 +53,31 @@
 ; The number of minute between sync intervals
 ;syncinterval = 60
 
-
 [Community Stat]
 
 ; Disables periodic community stat collection
 ;disablecommstat = false
 
 ; Collection interval for community stat
-;commstatinterval = 300
+;redditstatinterval = 60
 
 ; List of subreddit for community stat collection
-;subreddit = decred
+subreddit = decred
+subreddit = Dcrtrader
+subreddit = DCR
+
+; Number of minutes between Twitter stat collection
+twitterstatinterval = 1440
+; List of twitter handles community stat collection
+twitterhandle = decredproject
+
+; Number of minutes between Github stat collection
+githubstatinterval = 1440
+; List of Github repository to track
+githubrepository = decred/dcrd
+githubrepository = raedahgroup/dcrextdata
+
+; Number of hours between Youtube stat collection
+youtubestatinterval = 1440
+; Youtube data API key gotten from google developer console. https://developers.google.com/youtube/v3/getting-started
+  youtubedataapikey =
